Updated product_id field type
Original implementation represented product ID as an IntegerField.  Along with potential issues with leading zeros, databases such as Amazon's use alphanumeric product IDs.  A CharField seems to fill the role more effectively.  Also added .idea folder to the gitignore, which comes from setting it up as a project in PyCharm.