Fixed flaws in BB scraper tests
Incorrect use of self was causing tests to fail, fixed my interpretation of how that had to be used.  Also fixed some issues in Amazon tests, including the use of the "amzn.com/[ASID]" format.  I would swear this worked last night, but now it seems it needs to be "amzn.com/dp/[ASID]".  Testing/usage of the BB scraper seems to be very slow, with 7 tests running in roughly 60 seconds compared to 8 tests running in roughly 10.  This seems to be a result of the Chrome bootup time, not sure if this will be an issue in practice.