Custom site scraper fix, additional tests
Potential xpath bug averted, four more tests of the custom site scraper, and a beginning of tests for the RunScraper main method (still non-functional, probably need to move these to a separate file to get patches to work right).  Also made product->each change in scraper_service, plus a debug printout for running main in it.